requirements.txt
flask
flask-cors
openpyxl
pyinstaller
PyQt5
PyQtWebEngine
pymongo
python-dotenv
dnspython
